yuetransfer/requirements.txt
# Web framework
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.3.7

# SFTP and file handling
paramiko==3.3.1
cryptography==41.0.7
python-magic==0.4.27
Pillow==10.0.1
pydicom==2.4.3
nibabel==5.1.0

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1

# File processing and compression
zipfile36==0.1.3
rarfile==4.0

# System monitoring
psutil==5.9.5

# Configuration
PyYAML==6.0.1

# Production server
waitress==2.1.2

# Additional utilities
click==8.1.7
python-dateutil==2.8.2 